race-service renamed to race.service
